refactor: 기본 버튼 수정
