+++ src/app/styles/globals.css
@@ -49,3 +49,7 @@ input[type=number]::-webkit-out-spin-button {
 	-moz-appearance:none;
 	appearance:none;
 }
+
+.wrap{
+  @apply bg-white h-[100vh] relative p-[2rem];
+}
